Implement static client with most fields
For now, the conversion from a field to the static representation is
done at the passing of a parameter. This may cause some issues and needs
to be updated in the future.

Some fields were ignored due to the complexity of their format, they
will be implemented in a further review.